feat: add absolute position

+++ src/extension.ts
@@ -4,28 +4,57 @@ export const defaultIconPath = 'resources/broken.svg';
 export const defaultIconWidth = '14px';
 export const defaultIconHeight = '14px';
 
+function objectToCssString(settings: any): string {
+	let value = '';
+	const cssString = Object.keys(settings).map(setting => {
+		value = settings[setting];
+		if (typeof value === 'string' || typeof value === 'number') {
+			return `${setting}: ${value};`
+		}
+	}).join(' ');
+
+	return cssString;
+}
+
 export function activate(context: vscode.ExtensionContext) {
 	let diagnosticCollection = vscode.languages.createDiagnosticCollection('rust-panic-highlighter');
 	context.subscriptions.push(diagnosticCollection);
 
 	function createDecorationType(): vscode.TextEditorDecorationType {
-		let iconPathSetting = vscode.workspace.getConfiguration().get<string>('rustPanicHighlighter.iconPath') || '';
+		const isEnabled = vscode.workspace.getConfiguration().get<boolean>('icon.enabled') ?? true;
 
-		if (!iconPathSetting.endsWith('.svg')) {
-			vscode.window.showWarningMessage("Invalid icon path: Only SVG files are supported. Using default icon.");
+		if (!isEnabled) {
+			return vscode.window.createTextEditorDecorationType({});
+		}
+
+		let iconPathSetting = vscode.workspace.getConfiguration().get<string>('icon.path') || 'default';
+
+		if (iconPathSetting === "default") {
 			iconPathSetting = context.asAbsolutePath(defaultIconPath);
 		}
 
 		const iconPath = vscode.Uri.file(iconPathSetting);
-		const iconWidthSetting = vscode.workspace.getConfiguration().get<string>('rustPanicHighlighter.iconWidth') || defaultIconWidth;
-		const iconHeightSetting = vscode.workspace.getConfiguration().get<string>('rustPanicHighlighter.iconHeight') || defaultIconHeight;
+		const iconWidthSetting = vscode.workspace.getConfiguration().get<string>('icon.width') || defaultIconWidth;
+		const iconHeightSetting = vscode.workspace.getConfiguration().get<string>('icon.height') || defaultIconHeight;
+
+		const topValue = parseInt(iconWidthSetting) * 0.5;
+
+		const defaultCss = {
+			position: 'absolute',
+			top: `-${topValue}px`,
+			['z-index']: 1,
+			['pointer-events']: 'none',
+		};
+
+		const defaultCssString = objectToCssString(defaultCss);
 
 		return vscode.window.createTextEditorDecorationType({
 			after: {
 				contentIconPath: iconPath,
+				textDecoration: `none; ${defaultCssString}`,
 				width: iconWidthSetting,
 				height: iconHeightSetting,
-				margin: '0 10px',
+				margin: '0 1rem',
 			},
 		});
 	}
@@ -46,13 +75,16 @@ export function activate(context: vscode.ExtensionContext) {
 	});
 
 	vscode.workspace.onDidCloseTextDocument(doc => {
-		diagnosticCollection.delete(doc.uri);
+		if (doc.languageId === 'rust') {
+			diagnosticCollection.delete(doc.uri);
+		}
 	});
 
 	vscode.workspace.onDidChangeConfiguration(event => {
-		if (event.affectsConfiguration('rustPanicHighlighter.iconPath') ||
-			event.affectsConfiguration('rustPanicHighlighter.iconWidth') ||
-			event.affectsConfiguration('rustPanicHighlighter.iconHeight')) {
+		if (event.affectsConfiguration('icon.enabled') ||
+			event.affectsConfiguration('icon.path') ||
+			event.affectsConfiguration('icon.width') ||
+			event.affectsConfiguration('icon.height')) {
 
 			decorationType.dispose();
 			decorationType = createDecorationType();
@@ -72,7 +104,7 @@ function updateDiagnostics(doc: vscode.TextDocument, diagnosticCollection: vscod
 	let editor = vscode.window.activeTextEditor;
 	let rangesToDecorate: vscode.Range[] = [];
 
-	const severitySetting = vscode.workspace.getConfiguration().get<string>('rustPanicHighlighter.diagnosticSeverity');
+	const severitySetting = vscode.workspace.getConfiguration().get<string>('diagnostic.severity');
 	let severity: vscode.DiagnosticSeverity;
 
 	switch (severitySetting) {
@@ -98,34 +130,49 @@ function updateDiagnostics(doc: vscode.TextDocument, diagnosticCollection: vscod
 			continue;
 		}
 
-		if (line.text.includes("panic!") || line.text.includes("unwrap(") || line.text.includes("expect(")) {
+		if (line.text.includes("panic!(") || line.text.includes("unwrap()") || line.text.includes("expect(")) {
 			let diagnosticMessage = "";
 
-			if (line.text.includes("panic!")) {
+			if (line.text.includes("panic!(")) {
 				diagnosticMessage = "This line contains a 'panic!' which can cause a runtime panic.";
-			} else if (line.text.includes("unwrap(")) {
+			} else if (line.text.includes("unwrap()")) {
 				diagnosticMessage = "This line contains an 'unwrap()', which will panic if the result is None or Err.";
 			} else if (line.text.includes("expect(")) {
 				diagnosticMessage = "This line contains an 'expect()', which will panic if the result is None or Err.";
 			}
 
+			let startIndex = 0;
+			if (line.text.includes("panic!(")) {
+				startIndex = line.text.indexOf("panic!(");
+			} else if (line.text.includes("unwrap()")) {
+				startIndex = line.text.indexOf("unwrap()");
+			} else if (line.text.includes("expect(")) {
+				startIndex = line.text.indexOf("expect(");
+			}
+
+			const range = new vscode.Range(i, startIndex, i, line.text.length);
+			rangesToDecorate.push(range);
+
 			let diagnostic = new vscode.Diagnostic(
-				line.range,
+				range,
 				diagnosticMessage,
 				severity
 			);
 			diagnostics.push(diagnostic);
-
-			const range = new vscode.Range(i, line.text.length, i, line.text.length);
-			rangesToDecorate.push(range);
 		}
 	}
 
-	if (editor && rangesToDecorate.length > 0) {
-		editor.setDecorations(decorationType, rangesToDecorate);
+	if (editor) {
+		if (rangesToDecorate.length > 0) {
+			editor.setDecorations(decorationType, rangesToDecorate);
+		} else {
+			editor.setDecorations(decorationType, []);
+		}
 	}
 
-	if (diagnostics.length > 0) {
+	if (diagnostics.length === 0) {
+		diagnosticCollection.delete(doc.uri);
+	} else {
 		diagnosticCollection.set(doc.uri, diagnostics);
 	}
 }
